pyproject.toml
[project]
name = "medqa"
version = "1.0.0"
description = "Production medical question answering pipeline with intent classification, symptom extraction, RAG retrieval, and answer generation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = []

dependencies = [
    "groq>=0.9.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
    "tqdm>=4.66.0",
    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    # Frontend
    "streamlit>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
